requirements.txt
requests
flask
websocket-client
PILLOW
packaging
flask_sock